fix: allow returning `Response` objects from function Handlers
Per the [functions documentation][1], it seems like you should be able
to return a standard `Response` object from function handlers. However,
the existing types only allows a user to return a `HandlerResponse`
object.

This updates the Handler type to accept both types as a return value.

(It looks like https://github.com/netlify/functions/pull/351 also
addresses this issue, but that PR is a little stale at this point.)

[1]: https://docs.netlify.com/functions/get-started/?fn-language=ts#synchronous-function