feat: publish ESM version (#473)
**Which problem is this pull request solving?**

This module doesn't currently publish an ESM version. It's hard to use
this in ES modules, because it needs to be dynamically imported, but it
imports node builtins that can't be dynamically imported.

**Describe the solution you've chosen**

This PR uses `tsup` to build both cjs and esm. It also adds publint and
arethetypeswrong to check the output, and removes the one-line
`is-promise` CJS dependency.