fix: throw an error if we are passing an ambiguous set of options to the purge api
Prior to this commit, a call to `purgeCache` could supply any combination of siteID, siteSlug,
and domain.
If that does happen then the purge api endpoint as of today will do the following:

- If the request passed in SiteID and either SiteSlug or Domain, then the SiteID would be used
- If the request passed in SiteSlug and Domain, then the SiteSlug would be used

This is likely not what the user is wanting.
I've opened a pull-request in netlify-server to make it return an error response instead,
but I think we can also improve the situation in this package by throwing an earlier error,
one which they the user can spot when developing locally