fix: only purge alias by default if the purgeCache function is called within a deployed function (#576)
linear ticket: https://linear.app/netlify/issue/RUN-1417/

This patch will ensure that when `purgeCache` is called within a
deployed function, the purge will happen for the deployment which
contains the function. I.E. If it's the production deployment, it purges
production. If it's a branch deployment, it purges the branch. If it's a
deploy-preview, it purges the deploy-preview.